+++ sentry/src/main/java/io/sentry/adapters/OrientationSerializerAdapter.java
@@ -4,8 +4,8 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializationContext;
 import com.google.gson.JsonSerializer;
-import io.sentry.ILogger;
 import io.sentry.SentryLevel;
+import io.sentry.SentryOptions;
 import io.sentry.protocol.Device;
 import java.lang.reflect.Type;
 import java.util.Locale;
@@ -16,10 +16,10 @@ import org.jetbrains.annotations.NotNull;
 public final class OrientationSerializerAdapter
     implements JsonSerializer<Device.DeviceOrientation> {
 
-  private final @NotNull ILogger logger;
+  private final @NotNull SentryOptions options;
 
-  public OrientationSerializerAdapter(final @NotNull ILogger logger) {
-    this.logger = logger;
+  public OrientationSerializerAdapter(final @NotNull SentryOptions options) {
+    this.options = options;
   }
 
   @Override
@@ -28,7 +28,7 @@ public final class OrientationSerializerAdapter
     try {
       return src == null ? null : new JsonPrimitive(src.name().toLowerCase(Locale.ROOT));
     } catch (Exception e) {
-      logger.log(SentryLevel.ERROR, "Error when serializing DeviceOrientation", e);
+      options.getLogger().log(SentryLevel.ERROR, "Error when serializing DeviceOrientation", e);
     }
     return null;
   }