+++ sentry/src/main/java/io/sentry/adapters/SentryIdDeserializerAdapter.java
@@ -4,8 +4,8 @@ import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParseException;
-import io.sentry.ILogger;
 import io.sentry.SentryLevel;
+import io.sentry.SentryOptions;
 import io.sentry.protocol.SentryId;
 import java.lang.reflect.Type;
 import org.jetbrains.annotations.ApiStatus;
@@ -14,10 +14,10 @@ import org.jetbrains.annotations.NotNull;
 @ApiStatus.Internal
 public final class SentryIdDeserializerAdapter implements JsonDeserializer<SentryId> {
 
-  private final @NotNull ILogger logger;
+  private final @NotNull SentryOptions options;
 
-  public SentryIdDeserializerAdapter(final @NotNull ILogger logger) {
-    this.logger = logger;
+  public SentryIdDeserializerAdapter(final @NotNull SentryOptions options) {
+    this.options = options;
   }
 
   @Override
@@ -26,7 +26,7 @@ public final class SentryIdDeserializerAdapter implements JsonDeserializer<Sentr
     try {
       return json == null ? null : new SentryId(json.getAsString());
     } catch (Exception e) {
-      logger.log(SentryLevel.ERROR, "Error when deserializing SentryId", e);
+      options.getLogger().log(SentryLevel.ERROR, "Error when deserializing SentryId", e);
     }
     return null;
   }