+++ sentry/src/main/java/io/sentry/adapters/SpanIdSerializerAdapter.java
@@ -4,8 +4,8 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializationContext;
 import com.google.gson.JsonSerializer;
-import io.sentry.ILogger;
 import io.sentry.SentryLevel;
+import io.sentry.SentryOptions;
 import io.sentry.SpanId;
 import java.lang.reflect.Type;
 import org.jetbrains.annotations.ApiStatus;
@@ -14,10 +14,10 @@ import org.jetbrains.annotations.NotNull;
 @ApiStatus.Internal
 public final class SpanIdSerializerAdapter implements JsonSerializer<SpanId> {
 
-  private final @NotNull ILogger logger;
+  private final @NotNull SentryOptions options;
 
-  public SpanIdSerializerAdapter(final @NotNull ILogger logger) {
-    this.logger = logger;
+  public SpanIdSerializerAdapter(final @NotNull SentryOptions options) {
+    this.options = options;
   }
 
   @Override
@@ -25,7 +25,7 @@ public final class SpanIdSerializerAdapter implements JsonSerializer<SpanId> {
     try {
       return src == null ? null : new JsonPrimitive(src.toString());
     } catch (Exception e) {
-      logger.log(SentryLevel.ERROR, "Error when serializing SpanId", e);
+      options.getLogger().log(SentryLevel.ERROR, "Error when serializing SpanId", e);
     }
     return null;
   }