+++ sentry/src/main/java/io/sentry/adapters/SpanStatusDeserializerAdapter.java
@@ -4,8 +4,8 @@ import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParseException;
-import io.sentry.ILogger;
 import io.sentry.SentryLevel;
+import io.sentry.SentryOptions;
 import io.sentry.SpanStatus;
 import java.lang.reflect.Type;
 import java.util.Locale;
@@ -15,10 +15,10 @@ import org.jetbrains.annotations.NotNull;
 @ApiStatus.Internal
 public final class SpanStatusDeserializerAdapter implements JsonDeserializer<SpanStatus> {
 
-  private final @NotNull ILogger logger;
+  private final @NotNull SentryOptions options;
 
-  public SpanStatusDeserializerAdapter(final @NotNull ILogger logger) {
-    this.logger = logger;
+  public SpanStatusDeserializerAdapter(final @NotNull SentryOptions options) {
+    this.options = options;
   }
 
   @Override
@@ -27,7 +27,7 @@ public final class SpanStatusDeserializerAdapter implements JsonDeserializer<Spa
     try {
       return json == null ? null : SpanStatus.valueOf(json.getAsString().toUpperCase(Locale.ROOT));
     } catch (Exception e) {
-      logger.log(SentryLevel.ERROR, "Error when deserializing SpanStatus", e);
+      options.getLogger().log(SentryLevel.ERROR, "Error when deserializing SpanStatus", e);
     }
     return null;
   }