+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Feat: Add an overload for `startTransaction` that sets the created transaction to the Scope #1313
 * Ref: Separate user facing and protocol classes in the Performance feature (#1304)
 * Feat: Set SDK version on Transactions (#1307)
+* Fix: Use logger set on SentryOptions in GsonSerializer (#1308)
 
 # 4.3.0